✨ JE-36-feat: reviewtitle 컴포넌트 분리
